Uploaded BioImageArchive_templates_aggregated
BIA metadata template where attributes for specimen preparation, image acquisition and image analysis is concatenated in a single template. This is to reduce sample linkage in DataHub. 
User can choose the specific template based on lab practices. If specimens are always subjected to single measurement and then discarded, separate templates are not required. If a specimen is imaged multiple times (with same or different technique), sample splitting is required; meaning separate templates need to implemented.